hex: do not use `size_t` with negative integers

+++ include/ethc/hex.h
@@ -8,9 +8,9 @@ extern "C" {
 #include <stddef.h>
 #include <stdint.h>
 
-int eth_is_hex(const char *str, size_t len, int strict);
-int eth_hex_pad_left(char *rstr, const char *str, size_t len, size_t width);
-int eth_hex_pad_right(char *rstr, const char *str, size_t len, size_t width);
+int eth_is_hex(const char *str, int len, int strict);
+int eth_hex_pad_left(char *rstr, const char *str, int len, size_t width);
+int eth_hex_pad_right(char *rstr, const char *str, int len, size_t width);
 int eth_hex_from_bytes(char *rstr, const uint8_t *bytes, size_t len);
 
 #ifdef __cplusplus
